requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python==0.3.16
psutil
python-multipart
aiosqlite
numpy
jinja2
diskcache
